requirements.txt
requests
websocket-client
numpy
binance-api
datetime
telepot
tradingview_ta
continuous_threading
telethon